IMN-381 Move toTenantV1 in converterV1 file in tenant-readmodel-writer (#338)
